+++ cmd/main.go
@@ -357,6 +357,8 @@ func main() {
 			os.Exit(runDoctor(os.Args[2:]))
 		case "chat":
 			os.Exit(runChat(os.Args[2:]))
+		case "setup":
+			os.Exit(runSetup(os.Args[2:]))
 		case "version", "--version", "-v":
 			fmt.Println("CalvoProxy " + version)
 			return